Use FlashList for TransactionsFlatList
